USER BASE CONTACT [CREATE,READ,UPDATE,DELETE] COMPLETED

+++ controllers/contactsController.js
@@ -7,9 +7,13 @@ const Contacts = require('../models/contactModel');
 //@access   PRIVATE
 
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contacts.find({ user_id: req.user.id });
+  const contact = await Contacts.find({});
 
-  res.status(200).json({ success: true, data: contacts });
+  if (!contact) {
+    return res.status(404).json({ success: false, msg: 'Contact not found' });
+  }
+
+  res.status(200).json({ success: true, data: contact });
 });
 
 //@desc     Get single contact
@@ -17,13 +21,9 @@ const getContacts = asyncHandler(async (req, res) => {
 //@access   PRIVATE
 
 const getContact = asyncHandler(async (req, res) => {
-  const contact = await Contacts.findById(req.params.id);
+  const contacts = await Contacts.find({ user_id: req.user.id });
 
-  if (!contact) {
-    return res.status(404).json({ success: false, msg: 'Contact not found' });
-  }
-
-  res.status(200).json({ success: true, data: contact });
+  res.status(200).json({ success: true, data: contacts });
 });
 
 //@desc     Create new contact
@@ -59,6 +59,10 @@ const updateContact = asyncHandler(async (req, res) => {
     return res.status(404).json({ success: false, msg: 'Contact not found' });
   }
 
+  if (contact.user_id.toString() !== req.user.id) {
+    return res.status(401).json({ success: false, msg: 'Not authorized' });
+  }
+
   const updatedContact = await Contacts.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -81,6 +85,9 @@ const deleteContact = asyncHandler(async (req, res) => {
 
   if (!contact) {
     return res.status(404).json({ success: false, msg: 'Contact not found' });
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    return res.status(401).json({ success: false, msg: 'Not authorized' });
   }
 
   const deletedContact = await Contacts.findByIdAndDelete(req.params.id);
